test(jit-cache): decouple wheelhouse tooling
